requirements.txt
# Core
pandas>=2.0
numpy>=1.25
polars>=1.2.0
pyarrow>=15.0.0
duckdb>=1.0.0
pydantic>=2.5.0
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.66.0
requests>=2.32.0

# Modeling
scikit-learn>=1.4.0
lightgbm>=4.3.0
xgboost>=2.0.3
statsmodels>=0.14.0
optuna>=3.5.0

# Tracking & orchestration
mlflow>=2.12.0
prefect>=3.0.0

# App
streamlit>=1.36.0
plotly>=5.22.0

# Tests & dev
pytest>=8.0.0
pre-commit>=3.7.0
black>=24.3.0
ruff>=0.5.0
